got the 3d model to display inside of the corresponding route in react router dom
