+++ client/src/App.jsx
@@ -4,10 +4,8 @@ import { Outlet } from 'react-router-dom';
 import NavBar from './Components/Navbar/Navbar';
 import { ApolloClient, ApolloProvider, InMemoryCache, createHttpLink } from '@apollo/client';
 import { setContext } from '@apollo/client/link/context';
-import { Canvas } from '@react-three/fiber'
 // css
 import './styles/main.css';
-import Experience from './Components/Three/Experience';
 
 const httpLink = createHttpLink({
   uri: '/graphql',
@@ -24,15 +22,6 @@ const authLink = setContext((_, { headers }) => {
   };
 });
 
-function Light({ brightness, color }) {
-  <rectAreaLight
-    width={5}
-    height={5}
-    color={color}
-    intensity={brightness}
-  />
-}
-
 const client = new ApolloClient({
   link: authLink.concat(httpLink),
   cache: new InMemoryCache(),
@@ -44,10 +33,6 @@ function App() {
       <ApolloProvider client={client}>
         <NavBar />
         <Outlet />
-        {/* <Canvas>
-          <Light brightness={20} color={'#ff0000'} />
-          <Experience />
-        </Canvas> */}
       </ApolloProvider>
     </>
   )
