+++ client/src/main.jsx
@@ -10,18 +10,11 @@ import MyLibrary from './pages/MyLibrary.jsx';
 import Login from './pages/Login.jsx';
 import Signup from './pages/Signup.jsx';
 import SingleBook from './pages/SingleBook.jsx';
-import BookReader from './pages/BookReader.jsx'
-import { Canvas } from '@react-three/fiber';
-import Experience from './Components/Three/Experience.jsx';
+// import BookReader from './pages/BookReader.jsx'
 
-function Light({ brightness, color }) {
-  <rectAreaLight
-    width={5}
-    height={5}
-    color={color}
-    intensity={brightness}
-  />
-}
+import Three from './Components/Three/Three.jsx';
+
+
 // this establishes pages component structure and their paths
 const router = createBrowserRouter([
   {
@@ -51,20 +44,10 @@ const router = createBrowserRouter([
         path: '/singleBook/:id',
         element: <SingleBook />
       },
-      // {
-      //   path: '/bookReader/:bookId',
-      //   element: (
-      //     <>
-      //       <Canvas>
-
-      //         <Light brightness={20} color={'#ff0000'} />
-      //         <Experience />
-      //       </Canvas>
-      //     </>
-      //   )
-
-
-      // }
+      {
+        path: '/bookReader/:bookId',
+        element: <Three />
+      }
     ]
   }
 ]);
